hackrx_full_project/backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
faiss-cpu
pypdf
python-docx
extract-msg
groq
requests
